requirements.txt
Django==2.2.13
django-admin==2.0.1
django-admin-interface==0.12.2
django-admin-thumbnails==0.2.5
django-colorfield==0.3.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
model-mommy==1.3.2
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
screen==1.0.1
six==1.14.0
sqlparse==0.3.1
xlwt==1.3.0
